refactor: move form control state interface
